requirements.txt
typer>=0.12.0,<0.13.0
rich>=13.7.0,<14.0.0
pydantic>=2.6.0,<3.0.0
pyyaml>=6.0.0,<7.0.0
httpx>=0.27.0,<0.28.0
playwright>=1.44.0,<1.50.0
